requirements.txt
scipy
pandas
numpy
plotly
matplotlib
ffprobe
pydub
pillow
